Add option to print timestamp in session logs (#603)
Limitation: data arrives in chunks with escape sequences, sometimes the information is cut in odd places and only partial information arrives.  So eventually some timestamps will not end on a new line, but in between the text.   Or some escape sequences can not be cleaned correctly because the timestamp introduces patterns that were not in the original stream.